Created a CoordConverter
