+++ tablature_editor/src/Models/Coord/Coord.cs
