--- tablature_editor/src/Utils/Enums.cs
@@ -1,6 +1,0 @@
-﻿namespace TablatureEditor.Utils
-{
-    public enum WriteModes { Unity, Tenth, Twenyth, Thirtieth };
-    public enum SkipModes { Zero, One };
-    public enum CursorMovements { Left, Up, Right, Down, ExpandLeft, ExpandUp, ExpandRight, ExpandDown };
-}
